requirements.txt
django
django-widget-tweaks
django-crispy-forms
requests
jinja2
django-jinja
django-annoying
dj-database-url
dj-static
python-decouple
whitenoise
psycopg2
Pillow
Unipath
Markdown
bleach
django-mail-templated
chardet